Remove unecessary companion objects, rename method, move enum class
