+++ src/main/kotlin/com/amazon/opendistroforelasticsearch/indexstatemanagement/model/action/TransitionsActionConfig.kt
@@ -19,19 +19,14 @@ import com.amazon.opendistroforelasticsearch.indexstatemanagement.action.Action
 import com.amazon.opendistroforelasticsearch.indexstatemanagement.action.TransitionsAction
 import com.amazon.opendistroforelasticsearch.indexstatemanagement.model.ManagedIndexMetaData
 import com.amazon.opendistroforelasticsearch.indexstatemanagement.model.Transition
-import com.amazon.opendistroforelasticsearch.indexstatemanagement.util.ActionType
 import org.elasticsearch.client.Client
 import org.elasticsearch.cluster.service.ClusterService
 
-data class TransitionsActionConfig(val transitions: List<Transition>) : ActionConfig(TRANSITIONS_ACTION_TYPE, null, null) {
+data class TransitionsActionConfig(val transitions: List<Transition>) : ActionConfig(ActionType.TRANSITION, null, null) {
 
-    override fun toActionClass(
+    override fun toAction(
         clusterService: ClusterService,
         client: Client,
         managedIndexMetaData: ManagedIndexMetaData
     ): Action = TransitionsAction(clusterService, client, managedIndexMetaData, this)
-
-    companion object {
-        val TRANSITIONS_ACTION_TYPE = ActionType.TRANSITION
-    }
 }
